pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "figrecipe"
version = "0.20.0"
description = "Reproducible matplotlib wrapper with mm-precision layouts"
readme = "README.md"
license = "AGPL-3.0"
requires-python = ">=3.9"
authors = [
    { name = "Yusuke Watanabe", email = "ywatanabe@scitex.ai" },
]
keywords = [
    "matplotlib",
    "reproducibility",
    "visualization",
    "scientific",
    "plotting",
    "publication",
    "recipe",
    "yaml",
    "millimeter",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "ruamel.yaml>=0.17.0",
    "scipy>=1.7.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pre-commit>=3.5.0",
    "pyyaml>=6.0.0",
]
imaging = [
    "Pillow>=9.0.0",
]
seaborn = [
    "seaborn>=0.12.0",
    "pandas>=1.3.0",
]
editor = [
    "flask>=2.0.0",
    "Pillow>=9.0.0",
]
desktop = [
    "flask>=2.0.0",
    "Pillow>=9.0.0",
    "pywebview>=4.0.0",
    "qtpy>=2.0.0",
    "PyQt6>=6.0.0",
    "PyQt6-WebEngine>=6.0.0",
]
demo = [
    "playwright>=1.40.0",
    "pytesseract>=0.3.0",
]
graph = [
    "networkx>=3.0",
]
graph-interactive = [
    "networkx>=3.0",
    "pyvis>=0.3",
]
mcp = [
    "fastmcp>=2.0.0",
]
all = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pyyaml>=6.0.0",
    "Pillow>=9.0.0",
    "seaborn>=0.12.0",
    "pandas>=1.3.0",
    "flask>=2.0.0",
    "pywebview>=4.0.0",
    "playwright>=1.40.0",
    "pytesseract>=0.3.0",
    "networkx>=3.0",
    "pyvis>=0.3",
]

[project.scripts]
figrecipe = "figrecipe._cli:main"

[project.urls]
Homepage = "https://github.com/ywatanabe1989/figrecipe"
Documentation = "https://github.com/ywatanabe1989/figrecipe#readme"
Repository = "https://github.com/ywatanabe1989/figrecipe.git"
Issues = "https://github.com/ywatanabe1989/figrecipe/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/figrecipe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/figrecipe/__init__.py" = ["E402"]
